Made some significant changes to address the problems outlined in issue #80.
Primary change was to create all user repositories locally as soon as a user signs up.  I then refactored the Repo class and both the TaggingController and RepoController to find the local repository, this allowed me to remove most of the static methods in Repo.

Note: Added Repo#extract_info to create a new repo hash from a github repo.